requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
twilio==8.10.0
apscheduler==3.10.4
python-dotenv==1.0.0
fake-useragent==1.4.0
lxml==4.9.3
SQLAlchemy==2.0.23